runtime/common/io/requirements_updated.txt
# BioGPU Streaming Service - Python 3.12 Compatible Dependencies
# Updated: August 2025

# Core async libraries
# Note: asyncio is built-in to Python, no need to install separately
# aioredis is deprecated, use redis with async support
redis==5.0.8
kafka-python-ng==2.2.2
websockets==12.0
aiofiles==24.1.0

# Web framework and server
fastapi==0.111.1
uvicorn[standard]==0.30.5
python-multipart==0.0.9

# Database drivers
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
pymongo==4.8.0
motor==3.5.1  # Async MongoDB driver

# Authentication and security
PyJWT==2.9.0
cryptography==43.0.0
python-jose[cryptography]==3.3.0

# Data validation and serialization
pydantic==2.8.2
pydantic-settings==2.4.0

# HTTP client
httpx==0.27.0

# Data processing
numpy==2.0.1  # NumPy 2.0 is compatible with Python 3.12
pandas==2.2.2
h5py==3.11.0
pyarrow==17.0.0

# Logging and monitoring
structlog==24.4.0
python-json-logger==2.0.7
prometheus-client==0.20.0

# System monitoring
psutil==6.0.0
py-cpuinfo==9.0.0

# Container management
docker==7.1.0

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
pytest-mock==3.14.0
pytest-timeout==2.3.1

# Development tools
black==24.4.2
isort==5.13.2
flake8==7.1.0
mypy==1.11.1
pre-commit==3.7.1

# Additional async utilities
aiocache==0.12.2
aiohttp==3.10.2
tenacity==8.5.0  # For retry logic

# Message queue alternatives (optional)
# celery==5.4.0
# kombu==5.3.7

# Profiling and debugging (optional)
# py-spy==0.3.14
# memory-profiler==0.61.0